requirements.txt
# Telecom Room Monitor — Python Dependencies
# Install inside .venv:  pip install -r requirements.txt

adafruit-circuitpython-dht
adafruit-blinka
gpiozero
#lgpio
RPi.GPIO
RPLCD
smbus2
python-socketio[client]
